unified checks for replay writer and mettascope server
